remove django-simple-history
