+++ apps/betterangels-backend/betterangels_backend/settings.py
@@ -104,7 +104,6 @@ INSTALLED_APPS = [
     "post_office",
     "rest_framework",
     "organizations",
-    "simple_history",
     "strawberry_django",
     "pghistory",
     "pgtrigger",
@@ -125,7 +124,6 @@ MIDDLEWARE = [
     "django.contrib.auth.middleware.AuthenticationMiddleware",
     "django.contrib.messages.middleware.MessageMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
-    "simple_history.middleware.HistoryRequestMiddleware",
     "allauth.account.middleware.AccountMiddleware",
 ]
