+++ apps/betterangels-backend/pyproject.toml
@@ -34,7 +34,6 @@ hiredis = "^2.2.3"
 django-jazzmin = "^2.6.0"
 django-organizations = "^2.2.0"
 django-sesame = "^3.2.1"
-django-simple-history = "^3.4.0"
 strawberry-graphql-django = "^0.35.0"
 django-guardian = "^2.4.0"
 django-choices-field = "^2.2.2"
