+++ mypy.ini
@@ -65,9 +65,6 @@ ignore_missing_imports = True
 [mypy-sesame.*]
 ignore_missing_imports = True
 
-[mypy-simple_history.*]
-ignore_missing_imports = True
-
 [mypy-unittest_parametrize.*]
 ignore_missing_imports = True
